+++ packages/wagmi/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @web3modal/wagmi
 
+## 5.0.3
+
+### Patch Changes
+
+- fix: ethers5 coinbase issues. Turbo build issues. Upate cb connector.
+
+- Updated dependencies []:
+  - @web3modal/polyfills@5.0.3
+  - @web3modal/scaffold@5.0.3
+  - @web3modal/scaffold-react@5.0.3
+  - @web3modal/scaffold-utils@5.0.3
+  - @web3modal/scaffold-vue@5.0.3
+  - @web3modal/siwe@5.0.3
+
 ## 5.0.2
 
 ### Patch Changes